Initial implementation of neededness analysis
The analysis is done on the IR rather than the abstract assembly. The IR is
a little more difficult to work and I had to duplicate some of the work from
liveness analysis but ultimately it belongs in the IR. It currently generates
the set of needed variables now but doesn't do anything with that information
yet. I still need to make it delete lines that don't assign to needed
variables.

git-svn-id: https://cvs.concert.cs.cmu.edu/15411-f11/groups/eagle@1144 649c2953-11c0-447d-9ea0-900185aa3074