Starting to move some more logic from formatting to codegen.
* DIV/MOD both expanded much sooner
* SHL/RSH expanded sooner as well

This reduces some cases that we have to worry about in the formatter, although
it's possibly generating more code. We're still dealing with a lot of cases in
codegen, but with register coalescing we could worry about far fewer cases.

Allocation/liveness had to be updated to take raw operands (registers) into
account. Precoloring has become generic to deal with this and basically keep
track of live registers along with live temps.

Right now generate_seo is extremely slow for large programs and we fail two
tests:
  ../tests1/kite-return2.l1
  ../tests2/brachiosaurus-divide-zero.l1

Need to work on improving these

git-svn-id: https://cvs.concert.cs.cmu.edu/15411-f11/groups/eagle@1522 649c2953-11c0-447d-9ea0-900185aa3074