I fixed register coalescing. It works really well now.
The problem was that if you moved a temp into a register, the temp would
be used so it would be in the live set for that line. It also defines the
register though, so originally we would add the register to the live set
which added an interference between the two, so they could never be
coalesced. Now we don't add register defines to live sets if it's for a
MOV instruction. All unneeded callee saving is removed and some stupid
moves (like setting up parameters for functions) are optimized too. Look
at tests0/aneuploid.l3


git-svn-id: https://cvs.concert.cs.cmu.edu/15411-f11/groups/eagle@1909 649c2953-11c0-447d-9ea0-900185aa3074