Don't re-map reload locations
Add the notion of a tag which lives forever, this came up in cases like:

    spilled                       re-ssa'd
  spill t5                      spill t5
  reload t5 <- t5               reload t6 <- t5
  reload t5 <- t5               reload t7 <- t6

Where the problem is that t6 has never been spilled. Because we're in SSA form,
we only need to spill t5 once because its value never changes. Hence we just
need the last instruction to be 'reload t7 <- t5' and this commit changes the
notion of the reload source from a temp to a tag and tags never change