Massive amount of changes
Trying to use LinearMap everywhere instead of the garbage-collected HashMap.
This implies a much stronger notion of mutability pretty much everywhere.
Currently nothing compiles, and it gets the farthest compiling on the incoming
branch of rust itself.

Well isn't this just a giant clusterfuck caused by one tiny change of switching
maps. Either I need a better understanding of rust, or rust needs an easier way
to deal with mutable/immutable versions of the same structure.